update: coffer eid should be u64 type

+++ emodules/emod_manager/boot/boot.c
@@ -53,7 +53,7 @@ void emain_upper_half(
 	extern u8 _end; // defined in the linker script
 	paddr_t emod_manager_pa_end = (paddr_t)&_end;
 
-	extern int coffer_pid;
+	extern u64 coffer_pid;
 	coffer_pid = (int)eid;
 
 	/* upper half of enclave initialization */
